requirements-auth.txt
google-auth>=2.22.0 # maintained by Google: Snyk health 91
requests>=2.32.4 # exception: maintained by community, required by google-auth: Snyk health 91